+++ Memorize-Stanford-CS193p-2023/Cardify.swift
@@ -1,0 +1,37 @@
+//
+//  Cardify.swift
+//  Memorize-Stanford-CS193p-2023
+//
+//  Created by Serhii Lukashchuk on 2024-01-20.
+//
+
+import SwiftUI
+
+struct Cardify: ViewModifier {
+    var isFaceUp: Bool
+    
+    func body(content: Content) -> some View {
+        ZStack {
+            let base = RoundedRectangle(cornerRadius: Constants.cornerRadius)
+            Group {
+                base.fill(.white)
+                base.strokeBorder(lineWidth: Constants.lineWidth)
+                content
+            }
+            .opacity(isFaceUp ? 1 : 0)
+            base.fill()
+                .opacity(isFaceUp ? 0: 1)
+        }
+    }
+    
+    private struct Constants {
+        static let cornerRadius: CGFloat = 12
+        static let lineWidth: CGFloat = 2
+    }
+}
+
+extension View {
+    func cardify(isFaceUp: Bool) -> some View {
+        self.modifier(Cardify(isFaceUp: isFaceUp))
+    }
+}
